修复SPI和W25X16驱动
- PWM加入对非整数占空比的支持，占空比调节更加精密
- SPI Flash驱动加入对芯片忙状态的判断
- 修复SPI驱动
- 修复W25X16 SPI Flash驱动
- 新增SPI Flash JEDEC ID的读取，可用于判断芯片是否工作

+++ USER/pwm.h
@@ -20,6 +20,6 @@
 
 void PWM_Init(void);                                                    //初始化
 
-void PWM_SetDutyCycle(uint8_t dutyCycle);                               //设置占空比
+void PWM_SetDutyCycle(float dutyCycle);                               //设置占空比
 
 #endif
